feat(ie): encode complete usage reporting triggers
TS 29.244 clauses 8.2.19 and 8.2.41 define currently assigned Reporting Triggers and Usage Report Trigger bits across three octets. The old models used two and one octets respectively, placed EVETH and LIUSA in the wrong positions, and conflated quota holding time with quota exhaustion.

Represent the Release 18 triggers independently, including the distinct volume-quota and time-quota flags, and make quota-exhausted reports use VOLQU/TIMQU rather than the threshold triggers VOLTH/TIMTH.

Spec: 3GPP TS 29.244 V18.10.0, clauses 8.2.19 and 8.2.41, Figures 8.2.19-1 and 8.2.41-1.

Tests: reporting-trigger flag positions and Release 18 round trips; usage-report-trigger cross-octet round trip; quota report trigger assertions.

+++ src/ie/usage_report_smr.rs
@@ -204,7 +204,7 @@ mod tests {
         let max_report = UsageReport::new(
             UrrId::new(u32::MAX),
             SequenceNumber::new(u32::MAX),
-            UsageReportTrigger::new(u8::MAX),
+            UsageReportTrigger::new(u32::from(u8::MAX)),
         );
         let smr = UsageReportSmr::new(max_report);
         let marshaled = smr.marshal();
